requirements.txt
Django==1.6.1
South==0.8.4
anyjson==0.3.3
argparse==1.2.1
django-debug-toolbar==1.0.1
django-social-auth==0.7.28
django-taggit==0.11.2
djangorestframework==2.3.10
gevent==1.0
gevent-socketio==0.3
gevent-websocket==0.9.1
greenlet==0.4.2
httplib2==0.8
oauth2==1.5.211
psycopg2==2.5.2
python-openid==2.2.5
sqlparse==0.1.10
wsgiref==0.1.2
